+++ tests/unit/block/test_table.py
@@ -127,6 +127,38 @@ def test__aggregate_css_formatters_concatenation():
     assert res == 'style="aaaaa;bbbbb"'
 
 
+def test__jinja_calls_formatters_correctly():
+    df = pd.DataFrame(columns=['a', 'b', 'c'], index=['x', 'y', 'z'])
+
+    formatter = abtf.TableFormatter()
+    formatter.create_cell_level_css = MagicMock(side_effect=NotImplementedError)
+    table = abt.HTMLJinjaTableBlock(df, formatters=[formatter], use_default_formatters=False)
+
+    container = MagicMock()
+    actual_cfg = MagicMock()
+    table._write_contents(container, actual_cfg)
+
+    row_names = {args[0][0][1] for args in formatter.create_cell_level_css.call_args_list}
+    col_names = {args[0][0][2] for args in formatter.create_cell_level_css.call_args_list}
+    assert row_names == set(df.index) | {abtf.HEADER_ROW_NAME}
+    assert col_names == set(df.columns) | {abtf.INDEX_COL_NAME}
+
+
+def test__jinja_hides_multiindex_flattening():
+    df = pd.DataFrame(
+        columns=['a', 'b', 'c'],
+        index=pd.MultiIndex.from_tuples([('x', 'x'), ('y', 'y'), ('z', 'z')])
+    )
+
+    table = abt.HTMLJinjaTableBlock(df, formatters=[abtf.FmtExpandMultiIndex()], use_default_formatters=False)
+
+    container = MagicMock()
+    actual_cfg = MagicMock()
+    table._write_contents(container, actual_cfg)
+
+    assert "('x', 'x')" not in str(container.append.call_args_list[0][0][0])
+
+
 def test__get_header_iterable_multiindex():
     df = pd.DataFrame(np.arange(12, dtype=float).reshape(3, 4), index=['a', 'b', 'c'], columns=['aa', 'bb', 'cc', 'aa'])
     df['grouping'] = 'g'
@@ -134,8 +166,52 @@ def test__get_header_iterable_multiindex():
     df = df.groupby(['grouping', 'index']).sum().unstack()
     t = abt.HTMLJinjaTableBlock(df)
     result = t._get_header_iterable()
-    expected = [[['aa', 'aa', 'aa'], ['bb', 'bb', 'bb'], ['cc', 'cc', 'cc'], ['aa', 'aa', 'aa']],
-                ['a', 'b', 'c', 'a', 'b', 'c', 'a', 'b', 'c', 'a', 'b', 'c']]
+    expected = [
+        [
+            abt.IndexCell('aa', [('aa', 'a'), ('aa', 'b'), ('aa', 'c')], 3, 1),
+            abt.IndexCell('bb', [('bb', 'a'), ('bb', 'b'), ('bb', 'c')], 3, 1),
+            abt.IndexCell('cc', [('cc', 'a'), ('cc', 'b'), ('cc', 'c')], 3, 1),
+            abt.IndexCell('aa', [('aa', 'a'), ('aa', 'b'), ('aa', 'c')], 3, 1),
+        ],
+        [
+            abt.IndexCell('a', [('aa', 'a')], 1, 1),
+            abt.IndexCell('b', [('aa', 'b')], 1, 1),
+            abt.IndexCell('c', [('aa', 'c')], 1, 1),
+            abt.IndexCell('a', [('bb', 'a')], 1, 1),
+            abt.IndexCell('b', [('bb', 'b')], 1, 1),
+            abt.IndexCell('c', [('bb', 'c')], 1, 1),
+            abt.IndexCell('a', [('cc', 'a')], 1, 1),
+            abt.IndexCell('b', [('cc', 'b')], 1, 1),
+            abt.IndexCell('c', [('cc', 'c')], 1, 1),
+            abt.IndexCell('a', [('aa', 'a')], 1, 1),
+            abt.IndexCell('b', [('aa', 'b')], 1, 1),
+            abt.IndexCell('c', [('aa', 'c')], 1, 1),
+        ]
+    ]
+    assert result == expected
+
+
+def test__get_index_iterable_multiindex():
+    df = pd.DataFrame(np.arange(12, dtype=float).reshape(3, 4), index=['a', 'b', 'c'], columns=['aa', 'bb', 'cc', 'aa'])
+    df['grouping'] = 'g'
+    df = df.reset_index()
+    df = df.groupby(['grouping', 'index']).sum().unstack()
+    t = abt.HTMLJinjaTableBlock(df.T)
+    result = t._get_index_iterable()
+    expected = [
+        [abt.IndexCell('aa', [('aa', 'a'), ('aa', 'b'), ('aa', 'c')], 3, 1), abt.IndexCell('a', [('aa', 'a')], 1, 1),],
+        [abt.IndexCell('b', [('aa', 'b')], 1, 1)],
+        [abt.IndexCell('c', [('aa', 'c')], 1, 1)],
+        [abt.IndexCell('bb', [('bb', 'a'), ('bb', 'b'), ('bb', 'c')], 3, 1), abt.IndexCell('a', [('bb', 'a')], 1, 1),],
+        [abt.IndexCell('b', [('bb', 'b')], 1, 1)],
+        [abt.IndexCell('c', [('bb', 'c')], 1, 1)],
+        [abt.IndexCell('cc', [('cc', 'a'), ('cc', 'b'), ('cc', 'c')], 3, 1), abt.IndexCell('a', [('cc', 'a')], 1, 1),],
+        [abt.IndexCell('b', [('cc', 'b')], 1, 1)],
+        [abt.IndexCell('c', [('cc', 'c')], 1, 1)],
+        [abt.IndexCell('aa', [('aa', 'a'), ('aa', 'b'), ('aa', 'c')], 3, 1), abt.IndexCell('a', [('aa', 'a')], 1, 1),],
+        [abt.IndexCell('b', [('aa', 'b')], 1, 1)],
+        [abt.IndexCell('c', [('aa', 'c')], 1, 1)],
+    ]
     assert result == expected
 
 
@@ -144,4 +220,101 @@ def test__get_header_iterable_plain_index():
     p = pd.DataFrame(np.ones((4, 5)), columns=columns)
     t = abt.HTMLJinjaTableBlock(p)
     result = t._get_header_iterable()
-    assert result == [columns]
+    expected = [[
+        abt.IndexCell('a', ['a'], 1, 1),
+        abt.IndexCell('b', ['b'], 1, 1),
+        abt.IndexCell('c', ['c'], 1, 1),
+        abt.IndexCell('d', ['d'], 1, 1),
+        abt.IndexCell('e', ['e'], 1, 1),
+    ]]
+    assert result == expected
+
+
+def test__get_index_iterable_plain_index():
+    index = ['a', 'b', 'c', 'd', 'e']
+    p = pd.DataFrame(np.ones((5, 4)), index=index)
+    t = abt.HTMLJinjaTableBlock(p)
+    result = t._get_index_iterable()
+    expected = [
+        [abt.IndexCell('a', ['a'], 1, 1)],
+        [abt.IndexCell('b', ['b'], 1, 1)],
+        [abt.IndexCell('c', ['c'], 1, 1)],
+        [abt.IndexCell('d', ['d'], 1, 1)],
+        [abt.IndexCell('e', ['e'], 1, 1)],
+    ]
+    assert result == expected
+
+
+def test__header_depth_merge():
+    """
+    should coalesce vertically if merge_vertical is passed, except for the
+    bottom cell. for example, the following multiindex:
+    """
+    columns = pd.MultiIndex.from_tuples([('a', 'a', 'a')])
+    p = pd.DataFrame([], columns=columns)
+    t = abt.HTMLJinjaTableBlock(p, merge_vertical=True)
+    result = t._get_header_iterable()
+    expected = [
+        [abt.IndexCell('a', [('a', 'a', 'a')], 1, 2)],
+        [],
+        [abt.IndexCell('a', [('a', 'a', 'a')], 1, 1)],
+    ]
+    assert result == expected
+
+
+def test__header_depth_and_span_merge():
+    columns = pd.MultiIndex.from_tuples([('a', 'a', 'a'), ('a', 'a', 'a')])
+    p = pd.DataFrame([], columns=columns)
+    t = abt.HTMLJinjaTableBlock(p, merge_vertical=True)
+    result = t._get_header_iterable()
+    expected = [
+        [abt.IndexCell('a', [('a', 'a', 'a'), ('a', 'a', 'a')], 2, 2)],
+        [],
+        [
+            abt.IndexCell('a', [('a', 'a', 'a')], 1, 1),
+            abt.IndexCell('a', [('a', 'a', 'a')], 1, 1),
+        ],
+    ]
+    assert result == expected
+
+
+def test__no_depth_merge():
+    """
+    should coalesce vertically if merge_vertical is passed, except for the
+    bottom cell. for example, the following multiindex:
+    """
+    columns = pd.MultiIndex.from_tuples([('a', 'a', 'a')])
+    p = pd.DataFrame([], columns=columns)
+    t = abt.HTMLJinjaTableBlock(p)
+    result = t._get_header_iterable()
+    expected = [
+        [abt.IndexCell('a', [('a', 'a', 'a')], 1, 1)],
+        [abt.IndexCell('a', [('a', 'a', 'a')], 1, 1)],
+        [abt.IndexCell('a', [('a', 'a', 'a')], 1, 1)],
+    ]
+    assert result == expected
+
+
+def test__no_merge_if_parents_differ():
+    """
+    should not merge multiindex cells if their parent cells differ.
+    """
+    columns = pd.MultiIndex.from_tuples([('a', 'b', 'c'), ('c', 'b', 'a')])
+    p = pd.DataFrame([], columns=columns)
+    t = abt.HTMLJinjaTableBlock(p)
+    result = t._get_header_iterable()
+    expected = [
+        [
+            abt.IndexCell('a', [('a', 'b', 'c')], 1, 1),
+            abt.IndexCell('c', [('c', 'b', 'a')], 1, 1),
+        ],
+        [
+            abt.IndexCell('b', [('a', 'b', 'c')], 1, 1),
+            abt.IndexCell('b', [('c', 'b', 'a')], 1, 1),
+        ],
+        [
+            abt.IndexCell('c', [('a', 'b', 'c')], 1, 1),
+            abt.IndexCell('a', [('c', 'b', 'a')], 1, 1),
+        ],
+    ]
+    assert result == expected